% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSPCAShiny.R
\name{MDSPCAShiny}
\alias{MDSPCAShiny}
\title{Launch 'MDSPCAShiny' Interface}
\usage{
MDSPCAShiny()
}
\value{
Nothing
}
\description{
MDSPCAShiny() loads interactive user interface built using R 'shiny'.
}
\details{
Start MDSPCAShiny

The interactive user interface is to provide an easy way for multidimensional scaling and principal component analysis and downloading relevant plots.
}
\examples{
if(interactive()){
library(rmarkdown)
MDSPCAShiny()
}
}
\keyword{MDSPCAShiny}
